Add files via upload
SonifyPic is a groundbreaking project that explores the fusion of visual and auditory senses by converting images into immersive soundscapes. By leveraging the power of Optical Character Recognition (OCR) and Speech Synthesis, this Python-based project allows users to hear the audible representation of the content within images. Imagine hearing the rhythmic hum of a forest or the structured tones of a city skyline simply by analyzing an image.